# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PM App API",
        "description": "Automated Project Management Tool API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/risks/": {
            "get": {
                "tags": ["risks"],
                "summary": "Read Risks",
                "description": "Get all risks for a project.",
                "operationId": "read_risks_risks__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Project Id"
                        }
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RiskRead"
                                    },
                                    "title": "Response Read Risks Risks  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["risks"],
                "summary": "Create Risk",
                "description": "Create a new risk.",
                "operationId": "create_risk_risks__post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RiskCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RiskRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/risks/{risk_id}": {
            "get": {
                "tags": ["risks"],
                "summary": "Get Risk",
                "description": "Get a specific risk by ID.",
                "operationId": "get_risk_risks__risk_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "risk_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Risk Id"
                        }
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RiskRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["risks"],
                "summary": "Delete Risk",
                "description": "Delete a risk. Only Cogniters can delete risks.",
                "operationId": "delete_risk_risks__risk_id__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "risk_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Risk Id"
                        }
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/risks/{risk_id}/resolve": {
            "post": {
                "tags": ["risks"],
                "summary": "Resolve Risk",
                "description": "Resolve a risk with a decision record.\n\n- Only Cogniters can resolve risks\n- Status must be CLOSED or MITIGATED\n- Decision record is required",
                "operationId": "resolve_risk_risks__risk_id__resolve_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "risk_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Risk Id"
                        }
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RiskResolve"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RiskRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/risks/{risk_id}/reopen": {
            "post": {
                "tags": ["risks"],
                "summary": "Reopen Risk",
                "description": "Reopen a previously resolved risk.\n\n- Only Cogniters can reopen risks\n- Reason is required",
                "operationId": "reopen_risk_risks__risk_id__reopen_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "risk_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Risk Id"
                        }
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RiskReopen"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RiskRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/risks/{risk_id}/comments": {
            "get": {
                "tags": ["risks"],
                "summary": "Get Risk Comments",
                "description": "Get all comments for a risk.\n\nBoth Cogniters and assigned Clients can view comments.",
                "operationId": "get_risk_comments_risks__risk_id__comments_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "risk_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Risk Id"
                        }
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommentRead"
                                    },
                                    "title": "Response Get Risk Comments Risks  Risk Id  Comments Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["risks"],
                "summary": "Add Risk Comment",
                "description": "Add a comment to a risk.\n\nBoth Cogniters and assigned Clients can comment on risks.",
                "operationId": "add_risk_comment_risks__risk_id__comments_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "risk_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Risk Id"
                        }
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommentCreate": {
                "properties": {
                    "content": {
                        "type": "string",
                        "title": "Content"
                    }
                },
                "type": "object",
                "required": ["content"],
                "title": "CommentCreate",
                "description": "Schema for creating a comment."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "RiskCreate": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "category": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Category"
                    },
                    "impact_rationale": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Impact Rationale"
                    },
                    "date_identified": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Date Identified"
                    },
                    "probability": {
                        "$ref": "#/components/schemas/RiskProbability"
                    },
                    "impact": {
                        "$ref": "#/components/schemas/RiskImpact"
                    },
                    "status": {
                        "$ref": "#/components/schemas/RiskStatus",
                        "default": "Open"
                    },
                    "mitigation_plan": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Mitigation Plan"
                    },
                    "project_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Project Id"
                    }
                },
                "type": "object",
                "required": ["title", "description", "probability", "impact", "project_id"],
                "title": "RiskCreate",
                "description": "Schema for creating a risk."
            },
...
            "RiskProbability": {
                "type": "string",
                "enum": ["Low", "Medium", "High"],
                "title": "RiskProbability"
            },
            "RiskRead": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "project_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Project Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "probability": {
                        "$ref": "#/components/schemas/RiskProbability"
                    },
                    "impact": {
                        "$ref": "#/components/schemas/RiskImpact"
                    },
                    "status": {
                        "$ref": "#/components/schemas/RiskStatus"
                    },
                    "category": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Category"
                    },
                    "impact_rationale": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Impact Rationale"
                    },
                    "date_identified": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Date Identified"
                    },
                    "mitigation_plan": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Mitigation Plan"
                    },
                    "decision_record": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Decision Record"
                    },
                    "resolved_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Resolved At"
                    },
                    "resolved_by_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Resolved By Id"
                    },
                    "reopen_reason": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Reopen Reason"
                    },
                    "reopened_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Reopened At"
                    },
                    "reopened_by_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Reopened By Id"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "project_id",
                    "title",
                    "description",
                    "probability",
                    "impact",
                    "status",
                    "category",
                    "impact_rationale",
                    "date_identified",
                    "mitigation_plan",
                    "decision_record",
                    "resolved_at",
                    "resolved_by_id",
                    "reopen_reason",
                    "reopened_at",
                    "reopened_by_id"
                ],
                "title": "RiskRead",
                "description": "Schema for reading risk data.\n\nAll fields are explicitly defined to ensure OpenAPI required/nullable is accurate."
            },
            "RiskReopen": {
                "properties": {
                    "reason": {
                        "type": "string",
                        "title": "Reason"
                    }
                },
                "type": "object",
                "required": ["reason"],
                "title": "RiskReopen",
                "description": "Schema for reopening a resolved risk."
            },
            "RiskResolve": {
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/RiskStatus"
                    },
                    "decision_record": {
                        "type": "string",
                        "title": "Decision Record"
                    }
                },
                "type": "object",
                "required": ["status", "decision_record"],
                "title": "RiskResolve",
                "description": "Schema for resolving a risk."
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "auth/login"
                    }
                }
            }
        }
    }
}